Add implementation of Feature T839, Formatted cast of datetimes to/from character strings

+++ h2/src/main/org/h2/util/StringUtils.java
@@ -110,7 +110,7 @@ public class StringUtils {
     /**
      * Convert a string to a SQL literal. Null is converted to NULL. The text is
      * enclosed in single quotes. If there are any special characters, the
-     * method STRINGDECODE is used.
+     * Unicode character string literal is used.
      *
      * @param s the text to convert.
      * @return the SQL literal
@@ -1282,8 +1282,8 @@ public class StringUtils {
      * @param positiveValue the number to append
      * @return the specified string builder
      */
-    public static StringBuilder appendZeroPadded(StringBuilder builder, int length, long positiveValue) {
-        String s = Long.toString(positiveValue);
+    public static StringBuilder appendZeroPadded(StringBuilder builder, int length, int positiveValue) {
+        String s = Integer.toString(positiveValue);
         length -= s.length();
         for (; length > 0; length--) {
             builder.append('0');
